refactor: enhance trajectory dataclasses and restructure test modules
- Renamed trajectory-related test dataclasses to improve clarity and align with the new structure for consistency.
- Refactored mock data fixtures for pandas and ROSBag to standardize naming and provide better descriptions.
- Reorganized `mock` module and test paths to reflect the updated hierarchy for improved modularity and readability.

Issue TCT-52
