+++ tests/tests_dataclasses/tests_abstract_dataclasses/tests_abstract_trajectory_dataclass/__init__.py
@@ -1,0 +1,1 @@
+# coding=utf-8
